refactor: record DTO 분리
