+++ backend/src/main/java/com/ll/TeamProject/domain/user/dto/admin/LoginDto.java
@@ -1,4 +1,6 @@
-package com.ll.TeamProject.domain.user.dto;
+package com.ll.TeamProject.domain.user.dto.admin;
+
+import com.ll.TeamProject.domain.user.dto.UserDto;
 
 public record LoginDto(
         UserDto item,
