+++ backend/src/main/java/com/ll/TeamProject/domain/user/dto/admin/PasswordChangeRequest.java
@@ -1,0 +1,7 @@
+package com.ll.TeamProject.domain.user.dto.admin;
+
+import lombok.NonNull;
+
+public record PasswordChangeRequest(
+        @NonNull String password
+) {}
