+++ backend/src/main/java/com/ll/TeamProject/domain/user/dto/admin/UserLoginReqBody.java
@@ -1,0 +1,8 @@
+package com.ll.TeamProject.domain.user.dto.admin;
+
+import lombok.NonNull;
+
+public record UserLoginReqBody(
+        @NonNull String username,
+        @NonNull String password
+) { }
