+++ backend/src/main/java/com/ll/TeamProject/domain/user/dto/admin/VerificationCodeRequest.java
@@ -1,0 +1,8 @@
+package com.ll.TeamProject.domain.user.dto.admin;
+
+import lombok.NonNull;
+
+public record VerificationCodeRequest(
+        @NonNull String username,
+        @NonNull String email
+) { }
