+++ backend/src/main/java/com/ll/TeamProject/domain/user/dto/admin/VerificationCodeVerifyRequest.java
@@ -1,0 +1,8 @@
+package com.ll.TeamProject.domain.user.dto.admin;
+
+import lombok.NonNull;
+
+public record VerificationCodeVerifyRequest(
+        @NonNull String username,
+        @NonNull String verificationCode
+) { }
