+++ backend/src/main/java/com/ll/TeamProject/domain/user/repository/AuthenticationRepository.java
@@ -30,5 +30,4 @@ public interface AuthenticationRepository extends JpaRepository<Authentication, 
             @Param("startDate") LocalDateTime startDate,
             @Param("endDate") LocalDateTime endDate
     );
-
 }
